For 4281: removed QAB robot, updated UI tests
